// React/pu-project/src/Types/Get/CarMakeDto.ts
export type CarMakeDto = {
  id: number;
  name: string;
  fullName: string;
  imageUrl: string;
  description: string;
  foundedIn: string;
  foundedBy: string;
  headquarters: string;
  createdOn: string;
  createdBy: string;
  numberOfCars: number;
  lastModifiedOn: string;
  lastModifiedBy: string;
};